+++ src/CleanRegex/Internal/Prepared/Parser/Consumer/CommentConsumer.php
@@ -1,6 +1,7 @@
 <?php
 namespace TRegx\CleanRegex\Internal\Prepared\Parser\Consumer;
 
+use TRegx\CleanRegex\Internal\Prepared\Parser\Convention;
 use TRegx\CleanRegex\Internal\Prepared\Parser\Entity\Comment;
 use TRegx\CleanRegex\Internal\Prepared\Parser\EntitySequence;
 use TRegx\CleanRegex\Internal\Prepared\Parser\Feed\Feed;
@@ -8,6 +9,14 @@ use TRegx\CleanRegex\Internal\Prepared\Parser\Feed\StringConditions;
 
 class CommentConsumer implements Consumer
 {
+    /** @var Convention */
+    private $convention;
+
+    public function __construct(Convention $convention)
+    {
+        $this->convention = $convention;
+    }
+
     public function condition(Feed $feed): Condition
     {
         return new CommentCondition($feed);
@@ -22,7 +31,7 @@ class CommentConsumer implements Consumer
 
     private function commitToStrings(Feed $feed, StringConditions $strings): void
     {
-        $commentEnd = $feed->string("\n");
+        $commentEnd = $feed->oneOf($this->convention->lineEndings());
         while (!$feed->empty()) {
             if (!$commentEnd->consumable()) {
                 $strings->add($feed->letter());